Merged PR 4976: DateTimeTransformer implemented in ML.NET with assocated C++ wrapper
This is the ML.NET framework implementation to use the shared C++ library DateTimeTransformer. It includes the ML.NET C# code, as well as the C++ wrapper code.

It does NOT yet implement saving of the model or exporting to ONNX.

The C# code works and has its associated Unit Tests, but since we have not setup the new project in the ML.NET solution it cannot be built in the DataPipelines repo yet.

REVIEW: I am not sure how to make a CMAKE file for the C++ dll's, so that still needs to be added.