Merged PR 5026: Additional columns for DateTimeFeaturizer
Combined the functionality of both DateTimeFeaturizers by adding all the combined columns to this one.

Holiday information is still not in here as we are not sure how we are going to handle that yet.

I added 2 more libraries to the shared section. They are in the ORT repo and Dmitri said we are good to take a dependency on that.